[hw3] refactor: move API router to separate module
